add Currency to client utils
